requirements.txt
astpretty==1.2.1
mysqlclient==1.3.12
terminaltables==3.1.0
# WARNING replace dateutil!
python-dateutil==2.7.3
pynacl
